Stop using types to load rich values
The load method for a variable type will only ever be concerned with
loading values from ENV, which will either be nil or a string.

The dump method may have to handle many different incoming data types.
That's because the dump method handles the variable's value setter
method, which is also responsible for setting default values. So
any data type that could be passed in via the variable's setter or that
could be specified as a default value should be handled by the dump
method, which dumps it into ENV. The load method loads it beack from
ENV.