Back to providing specific versions for RVM on Travis
Updating RVM before_install is too late in the process. The requested
RVM Ruby version is installed before that hook is run.

An alternative would be not to specify RVM versions as part of the
build matrix but to set RVM_RUBY_VERSION environment variables to
control the build matrix and to use that variable in before_install to
manually install/use the requested Ruby after upgrading RVM. That's for
another day, if trying to keep JRuby and Rubinius versions up to date
becomes painful.